refactor[#554] : modal.ts 파일로 useModal 이동
